test(sdk): Move tests into their correct module.
