+++ crates/matrix-sdk/src/event_cache/room/mod.rs
@@ -555,3 +555,293 @@ impl RoomEventCacheState {
         self.waited_for_initial_prev_token = false;
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use matrix_sdk_common::deserialized_responses::SyncTimelineEvent;
+    use matrix_sdk_test::async_test;
+    use ruma::{
+        event_id,
+        events::{relation::RelationType, room::message::RoomMessageEventContentWithoutRelation},
+        room_id, user_id, RoomId,
+    };
+
+    use crate::test_utils::{events::EventFactory, logged_in_client};
+
+    #[async_test]
+    async fn test_event_with_redaction_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.text_msg("Original event").event_id(original_id).into(),
+            f.redaction(original_id).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_edit_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.text_msg("Original event").event_id(original_id).into(),
+            f.text_msg("* An edited event")
+                .edit(
+                    original_id,
+                    RoomMessageEventContentWithoutRelation::text_plain("And edited event"),
+                )
+                .event_id(related_id)
+                .into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_reply_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.text_msg("Original event").event_id(original_id).into(),
+            f.text_msg("A reply").reply_to(original_id).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_thread_reply_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.text_msg("Original event").event_id(original_id).into(),
+            f.text_msg("A reply").in_thread(original_id, related_id).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_reaction_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.text_msg("Original event").event_id(original_id).into(),
+            f.reaction(original_id, ":D".to_owned()).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_poll_response_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.poll_start("Poll start event", "A poll question", vec!["An answer"])
+                .event_id(original_id)
+                .into(),
+            f.poll_response("1", original_id).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_poll_end_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let f = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        assert_relations(
+            room_id,
+            f.poll_start("Poll start event", "A poll question", vec!["An answer"])
+                .event_id(original_id)
+                .into(),
+            f.poll_end("Poll ended", original_id).event_id(related_id).into(),
+            f,
+        )
+        .await;
+    }
+
+    #[async_test]
+    async fn test_event_with_filtered_relationships() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let associated_related_id = event_id!("$recursive_related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let event_factory = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        let original_event = event_factory.text_msg("Original event").event_id(original_id).into();
+        let related_event = event_factory
+            .text_msg("* Edited event")
+            .edit(original_id, RoomMessageEventContentWithoutRelation::text_plain("Edited event"))
+            .event_id(related_id)
+            .into();
+        let associated_related_event =
+            event_factory.redaction(related_id).event_id(associated_related_id).into();
+
+        let client = logged_in_client(None).await;
+
+        let event_cache = client.event_cache();
+        event_cache.subscribe().unwrap();
+
+        client.base_client().get_or_create_room(room_id, matrix_sdk_base::RoomState::Joined);
+        let room = client.get_room(room_id).unwrap();
+
+        let (room_event_cache, _drop_handles) = room.event_cache().await.unwrap();
+
+        // Save the original event.
+        room_event_cache.save_event(original_event).await;
+
+        // Save the related event.
+        room_event_cache.save_event(related_event).await;
+
+        // Save the associated related event, which redacts the related event.
+        room_event_cache.save_event(associated_related_event).await;
+
+        let filter = Some(vec![RelationType::Replacement]);
+        let (event, related_events) =
+            room_event_cache.event_with_relations(original_id, filter).await.unwrap();
+        // Fetched event is the right one.
+        let cached_event_id = event.event_id().unwrap();
+        assert_eq!(cached_event_id, original_id);
+
+        // There are both the related id and the associatively related id
+        assert_eq!(related_events.len(), 2);
+
+        let related_event_id = related_events[0].event_id().unwrap();
+        assert_eq!(related_event_id, related_id);
+        let related_event_id = related_events[1].event_id().unwrap();
+        assert_eq!(related_event_id, associated_related_id);
+
+        // Now we'll filter threads instead, there should be no related events
+        let filter = Some(vec![RelationType::Thread]);
+        let (event, related_events) =
+            room_event_cache.event_with_relations(original_id, filter).await.unwrap();
+        // Fetched event is the right one.
+        let cached_event_id = event.event_id().unwrap();
+        assert_eq!(cached_event_id, original_id);
+        // No Thread related events found
+        assert!(related_events.is_empty());
+    }
+
+    #[async_test]
+    async fn test_event_with_recursive_relation() {
+        let original_id = event_id!("$original");
+        let related_id = event_id!("$related");
+        let associated_related_id = event_id!("$recursive_related");
+        let room_id = room_id!("!galette:saucisse.bzh");
+        let event_factory = EventFactory::new().room(room_id).sender(user_id!("@ben:saucisse.bzh"));
+
+        let original_event = event_factory.text_msg("Original event").event_id(original_id).into();
+        let related_event = event_factory
+            .text_msg("* Edited event")
+            .edit(original_id, RoomMessageEventContentWithoutRelation::text_plain("Edited event"))
+            .event_id(related_id)
+            .into();
+        let associated_related_event =
+            event_factory.redaction(related_id).event_id(associated_related_id).into();
+
+        let client = logged_in_client(None).await;
+
+        let event_cache = client.event_cache();
+        event_cache.subscribe().unwrap();
+
+        client.base_client().get_or_create_room(room_id, matrix_sdk_base::RoomState::Joined);
+        let room = client.get_room(room_id).unwrap();
+
+        let (room_event_cache, _drop_handles) = room.event_cache().await.unwrap();
+
+        // Save the original event.
+        room_event_cache.save_event(original_event).await;
+
+        // Save the related event.
+        room_event_cache.save_event(related_event).await;
+
+        // Save the associated related event, which redacts the related event.
+        room_event_cache.save_event(associated_related_event).await;
+
+        let (event, related_events) =
+            room_event_cache.event_with_relations(original_id, None).await.unwrap();
+        // Fetched event is the right one.
+        let cached_event_id = event.event_id().unwrap();
+        assert_eq!(cached_event_id, original_id);
+
+        // There are both the related id and the associatively related id
+        assert_eq!(related_events.len(), 2);
+
+        let related_event_id = related_events[0].event_id().unwrap();
+        assert_eq!(related_event_id, related_id);
+        let related_event_id = related_events[1].event_id().unwrap();
+        assert_eq!(related_event_id, associated_related_id);
+    }
+
+    async fn assert_relations(
+        room_id: &RoomId,
+        original_event: SyncTimelineEvent,
+        related_event: SyncTimelineEvent,
+        event_factory: EventFactory,
+    ) {
+        let client = logged_in_client(None).await;
+
+        let event_cache = client.event_cache();
+        event_cache.subscribe().unwrap();
+
+        client.base_client().get_or_create_room(room_id, matrix_sdk_base::RoomState::Joined);
+        let room = client.get_room(room_id).unwrap();
+
+        let (room_event_cache, _drop_handles) = room.event_cache().await.unwrap();
+
+        // Save the original event.
+        let original_event_id = original_event.event_id().unwrap();
+        room_event_cache.save_event(original_event).await;
+
+        // Save an unrelated event to check it's not in the related events list.
+        let unrelated_id = event_id!("$2");
+        room_event_cache
+            .save_event(event_factory.text_msg("An unrelated event").event_id(unrelated_id).into())
+            .await;
+
+        // Save the related event.
+        let related_id = related_event.event_id().unwrap();
+        room_event_cache.save_event(related_event).await;
+
+        let (event, related_events) =
+            room_event_cache.event_with_relations(&original_event_id, None).await.unwrap();
+        // Fetched event is the right one.
+        let cached_event_id = event.event_id().unwrap();
+        assert_eq!(cached_event_id, original_event_id);
+
+        // There is only the actually related event in the related ones
+        assert_eq!(related_events.len(), 1);
+        let related_event_id = related_events[0].event_id().unwrap();
+        assert_eq!(related_event_id, related_id);
+    }
+}
